requirements.txt
prefect>=3.0,<4.0
anthropic>=0.39.0
psycopg2-binary>=2.9
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
pydantic>=2.0
pgvector>=0.3.0
python-dotenv>=1.0
arrow>=1.3
rich>=13.0
pdfplumber>=0.11
pandas>=2.0
openpyxl>=3.1
xlrd>=2.0
